websocket-chatroom/backend/requirements.txt
# FastAPI核心
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.19.0

# Redis（分布式模式）
redis==5.0.1

# JWT认证
python-jose[cryptography]==3.3.0

# 配置管理
pydantic-settings==2.1.0

# WebSocket
websockets==12.0
